make debugging on windows working again, complete replace mpfr with preprocessor makro, don't use openssl in debug build

+++ src/lib/DecayDecimal.cpp
@@ -8,8 +8,8 @@ DecayDecimal::~DecayDecimal()
 
 void DecayDecimal::applyDecay(Duration duration)
 {
-	Decimal durationSeconds(static_cast<unsigned int>(std::chrono::duration_cast<std::chrono::seconds>(duration).count()));
-	Decimal secondsPerYear(MAGIC_NUMBER_GREGORIAN_CALENDER_SECONDS_PER_YEAR);
+	Decimal durationSeconds(static_cast<long>(std::chrono::duration_cast<std::chrono::seconds>(duration).count()));
+	Decimal secondsPerYear(static_cast<long>(MAGIC_NUMBER_GREGORIAN_CALENDER_SECONDS_PER_YEAR));
 	Decimal timeFactor = (durationSeconds / secondsPerYear);
 	timeFactor = 2 ^ timeFactor;
 	(*this) /= timeFactor;
